impl extractConflictDataFromErrorObj
